authentication flow fixes

+++ lib/tasktracker_web/controllers/user_controller.ex
@@ -4,7 +4,7 @@ defmodule TasktrackerWeb.UserController do
   alias Tasktracker.Accounts
   alias Tasktracker.Accounts.User
 
-  plug :authenticate_user when action not in [:new]
+  plug :authenticate_user when action not in [:new, :create]
 
   def index(conn, _params) do
     users = Accounts.list_users()
@@ -17,11 +17,15 @@ defmodule TasktrackerWeb.UserController do
     render(conn, "new.html", changeset: changeset)
   end
 
+  # TODO: Refactor this
   def create(conn, %{"user" => user_params}) do
     case Accounts.create_user(user_params) do
       {:ok, user} ->
         conn
-        |> put_flash(:info, "User created successfully.")
+        |> put_session(:logged_in_as, user.name)
+        |> put_session(:user_id, user.id)
+        |> put_session(:logged_in, true)
+        |> put_flash(:info, "Registration succesful!")
         |> redirect(to: user_path(conn, :show, user))
       {:error, %Ecto.Changeset{} = changeset} ->
         render(conn, "new.html", changeset: changeset)
@@ -53,11 +57,20 @@ defmodule TasktrackerWeb.UserController do
   end
 
   def delete(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
-    {:ok, _user} = Accounts.delete_user(user)
+    my_id = get_session(conn, :user_id)
+    
+    if (id == Integer.to_string(my_id)) do
+      conn
+      |> put_flash(:error, "Can't delete your own account.")
+      |> redirect(to: user_path(conn, :index))
+      |> halt()
+    else 
+      user = Accounts.get_user!(id)
+      {:ok, _user} = Accounts.delete_user(user)
+      conn
+      |> put_flash(:info, "User deleted successfully.")
+      |> redirect(to: user_path(conn, :index))
+    end
 
-    conn
-    |> put_flash(:info, "User deleted successfully.")
-    |> redirect(to: user_path(conn, :index))
   end
 end
